Almost finished, just need to add visibility state for when no sponsors
