+++ client/src/components/navBar/NavBar.js
@@ -16,6 +16,7 @@ var NavBar = ({visibleBlock}) => {
           <NavBarEntry visibleBlock={visibleBlock} blockId="landing-block" text="Home" />
           <NavBarEntry visibleBlock={visibleBlock} blockId="eventdate-block" text="Dates" />
           <NavBarEntry visibleBlock={visibleBlock} blockId="about-us-block" text="About" />
+          <NavBarEntry visibleBlock={visibleBlock} blockId="sponsorBlock" text="Sponsors" />
           <NavBarEntry visibleBlock={visibleBlock} blockId="faq-block" text="FAQ" />
         </ul>
       </div>
