+++ client/src/components/sponsorsBlock/TEMPLATEsponsorInfo.json
@@ -1,0 +1,24 @@
+{
+    "sponsors": [
+      {
+        "companyName": "Tech Innovators Inc.",
+        "tier": "Title",
+        "industry": "Software Development",
+        "description": "Tech Innovators Inc. is a leading software development company specializing in innovative solutions for the tech industry. We are committed to fostering growth and providing cutting-edge technologies to our clients.",
+        "logo":"/sponsorLogos/[filename]",
+        //Everything below this point is optional.
+        "contactPerson": {
+          "name": "Jane Smith",
+          "email": "jane.smith@techinnovators.com",
+          "phone": "+1-800-123-4567"
+        },
+        "boothNumber": "A1",
+        "website": "https://www.techinnovators.com",
+        "socialMedia": {
+          "linkedin": "https://www.linkedin.com/company/tech-innovators",
+          "twitter": "https://twitter.com/techinnovators"
+        }
+      }
+    ]
+  }
+
