+++ client/src/components/sponsorsBlock/sponsorInfo.json
@@ -1,8 +1,8 @@
 {
     "sponsors": [
       {
-        "companyName": "Tech Innovators Inc.",
-        "tier": "Platinum",
+        "name": "Tech Innovators Inc.",
+        "tier": "Title",
         "industry": "Software Development",
         "contactPerson": {
           "name": "Jane Smith",
@@ -15,10 +15,11 @@
           "linkedin": "https://www.linkedin.com/company/tech-innovators",
           "twitter": "https://twitter.com/techinnovators"
         },
+        "logo": "/sponsorLogos/sample.jpg",
         "description": "Tech Innovators Inc. is a leading software development company specializing in innovative solutions for the tech industry. We are committed to fostering growth and providing cutting-edge technologies to our clients."
       },
       {
-        "companyName": "Green Energy Solutions",
+        "name": "Green Energy Solutions",
         "tier": "Gold",
         "industry": "Renewable Energy",
         "contactPerson": {
@@ -32,10 +33,11 @@
           "linkedin": "https://www.linkedin.com/company/green-energy-solutions",
           "twitter": "https://twitter.com/greenenergysol"
         },
+        "logo":"/compE_purple.svg",
         "description": "Green Energy Solutions is at the forefront of renewable energy technology, providing sustainable solutions to power the future. Join us in our mission to create a greener planet."
       },
       {
-        "companyName": "Future Tech Labs",
+        "name": "Future Tech Labs",
         "tier": "Silver",
         "industry": "Artificial Intelligence",
         "contactPerson": {
@@ -49,11 +51,12 @@
           "linkedin": "https://www.linkedin.com/company/future-tech-labs",
           "twitter": "https://twitter.com/futuretechlabs"
         },
+        "logo":"/logo512.png",
         "description": "Future Tech Labs is dedicated to advancing artificial intelligence and machine learning. Our innovations drive the future of technology, making the impossible possible."
       },
       {
-        "companyName": "NextGen Robotics",
-        "tier": "Bronze",
+        "name": "NextGen Robotics",
+        "tier": "Silver",
         "industry": "Robotics",
         "contactPerson": {
           "name": "Bob Brown",
@@ -66,7 +69,14 @@
           "linkedin": "https://www.linkedin.com/company/nextgen-robotics",
           "twitter": "https://twitter.com/nextgenrobotics"
         },
+        "logo":"https://c8.alamy.com/comp/2AH6RFF/real-estate-company-logo-design-template-blue-house-and-building-concept-construction-architecture-element-apartment-condo-rouded-window-shape-2AH6RFF.jpg",
         "description": "NextGen Robotics is revolutionizing the robotics industry with our state-of-the-art technology and innovative solutions. Explore our cutting-edge products and see the future of robotics today."
+      },
+      {
+        "name":"test",
+        "tier":"Bronze",
+        "logo":"https://c8.alamy.com/comp/2AH6RFF/real-estate-company-logo-design-template-blue-house-and-building-concept-construction-architecture-element-apartment-condo-rouded-window-shape-2AH6RFF.jpg",
+        "description":"test test test"
       }
     ]
   }
